Bugfixes (#45)
* Fix bearer token missing in failure responses

* Fix empty check in UnmarshalPrevious

* Fix generated types in resource model

+++ python/rpdk/go/resolver.py
@@ -2,9 +2,9 @@ from rpdk.core.jsonutils.resolver import UNDEFINED, ContainerType
 
 PRIMITIVE_TYPES = {
     "string": "string",
-    "integer": "int",
+    "integer": "int64",
     "boolean": "bool",
-    "number": "int64",
+    "number": "float64",
     UNDEFINED: "Type",
 }
